+++ datadog/data_source_datadog_ip_ranges.go
@@ -1,7 +1,7 @@
 package datadog
 
 import (
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	datadog "github.com/zorkian/go-datadog-api"
 )