+++ datadog/data_source_datadog_ip_ranges_test.go
@@ -3,7 +3,7 @@ package datadog
 import (
 	"testing"
 
-	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 )
 
 func TestAccDatadogIpRangesDatasource_existing(t *testing.T) {