+++ datadog/import_datadog_logs_pipeline_test.go
@@ -1,7 +1,7 @@
 package datadog
 
 import (
-	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"testing"
 )