+++ datadog/resource_datadog_monitor.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/zorkian/go-datadog-api"
 )
 
@@ -136,6 +136,15 @@ func resourceDatadogMonitor() *schema.Resource {
 				Type:     schema.TypeInt,
 				Optional: true,
 				Default:  10,
+				DiffSuppressFunc: func(k, oldVal, newVal string, d *schema.ResourceData) bool {
+					if !d.Get("notify_no_data").(bool) {
+						if newVal != oldVal {
+							log.Printf("[DEBUG] Ignore the no_data_timeframe change of monitor '%s' because notify_no_data is false.", d.Get("name"))
+						}
+						return true
+					}
+					return newVal == oldVal
+				},
 			},
 			"renotify_interval": {
 				Type:     schema.TypeInt,